feat: Implement PlaylistGenie homepage and refactor creation flow
- Creates a new homepage at `/` with PlaylistGenie branding and marketing copy.
- Includes a call-to-action button on the homepage that directs you to the playlist creation page.
- Moves the existing playlist generation functionality to a new route `/create`.
- Updates application metadata (title and description) to reflect the PlaylistGenie app.
- Styles the homepage using Tailwind CSS to evoke a retro, zine-like feel.
